summary page inside app
