+++ apps/festival/festival/src/app/dashboard/tunnel/movie-tunnel-routing.module.ts
@@ -85,7 +85,7 @@ const tunnelRoutes: Routes = [
       },
       {
         path: 'summary',
-        loadChildren: () => import('@blockframes/movie/pages/summary/summary.module').then(m => m.TunnelSummaryModule)
+        loadChildren: () => import('./summary/summary.module').then(m => m.TunnelSummaryModule)
       },
       // Last page
       {
