+++ apps/festival/festival/src/app/dashboard/tunnel/summary/summary.component.ts
@@ -1,13 +1,14 @@
-import { Component, ChangeDetectionStrategy } from '@angular/core';
+import { Component, ChangeDetectionStrategy, OnInit, OnDestroy } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MovieService, MovieQuery, Movie } from '@blockframes/movie/+state';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { MovieTunnelComponent } from '../movie-tunnel.component';
 import { FormGroup, FormArray } from '@angular/forms';
 import { getMoviePublishStatus, getCurrentApp } from '@blockframes/utils/apps';
 import { RouterQuery } from '@datorama/akita-ng-router-store';
 import { mergeDeep } from '@blockframes/utils/helpers';
 import { map } from 'rxjs/operators';
+import { MovieFormShellComponent } from '@blockframes/movie/pages/shell/shell.component';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'festival-summary-tunnel',
@@ -15,14 +16,17 @@ import { map } from 'rxjs/operators';
   styleUrls: ['./summary.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class TunnelSummaryComponent {
-  form = this.tunnel.form;
-  isPublished$ = this.query.selectActive(movie => movie.main.storeConfig.status).pipe(
+export class TunnelSummaryComponent implements OnInit, OnDestroy{
+  form = this.shell.form;
+  subscription: Subscription;
+  missingFields: string[] = [];
+  invalidFields: string[] = [];
+  isPublished$ = this.query.selectActive(movie => movie.storeConfig.status).pipe(
     map(status => status === 'accepted' || status === 'submitted')
   )
 
   constructor(
-    private tunnel: MovieTunnelComponent,
+    private shell: MovieFormShellComponent,
     private router: Router,
     private route: ActivatedRoute,
     private service: MovieService,
@@ -30,6 +34,15 @@ export class TunnelSummaryComponent {
     private snackBar: MatSnackBar,
     private routerQuery: RouterQuery
   ) { }
+
+  ngOnInit(): void {
+    this.findInvalidControlsRecursive(this.form);
+    this.subscription = this.form.valueChanges.subscribe(() => this.findInvalidControlsRecursive(this.form));
+  }
+
+  ngOnDestroy(): void {
+    this.subscription.unsubscribe();
+  }
 
   public getPath(segment: string) {
     const { movieId } = this.route.snapshot.params;
@@ -40,8 +53,8 @@ export class TunnelSummaryComponent {
     if (this.form.valid) {
       const movie: Movie = mergeDeep(this.query.getActive(), this.form.value);
       const currentApp = getCurrentApp(this.routerQuery);
-      movie.main.storeConfig.status = getMoviePublishStatus(currentApp); // @TODO (#2765)
-      movie.main.storeConfig.appAccess.festival = true;
+      movie.storeConfig.status = getMoviePublishStatus(currentApp); // @TODO (#2765)
+      movie.storeConfig.appAccess.festival = true;
       await this.service.update(movie.id, movie);
       this.form.markAsPristine();
       const ref = this.snackBar.open('Movie Online !!', '', { duration: 1000 });
@@ -51,28 +64,28 @@ export class TunnelSummaryComponent {
       })
     } else {
       // Log the invalid forms
-      /* console.error(this.findInvalidControlsRecursive(this.form)) */
       this.snackBar.open('Fill all mandatory fields before submitting', '', { duration: 2000 });
     }
   }
 
   /* Utils function to get the list of invalid form. Not used yet, but could be useful later */
-  public findInvalidControlsRecursive(formToInvestigate: FormGroup | FormArray): string[] {
-    const invalidControls: string[] = [];
+  public findInvalidControlsRecursive(formToInvestigate: FormGroup | FormArray) {
+    this.invalidFields = [];
+    this.missingFields = [];
     const recursiveFunc = (form: FormGroup | FormArray) => {
       Object.keys(form.controls).forEach(field => {
         const control = form.get(field);
         if (control.invalid) {
-          invalidControls.push(field);
+          this.invalidFields.push(field);
         }
-        if (control instanceof FormGroup) {
-          recursiveFunc(control);
-        } else if (control instanceof FormArray) {
+        if (!control.value) {
+          this.missingFields.push(field);
+        }
+        if (control instanceof FormArray || control instanceof FormGroup) {
           recursiveFunc(control);
         }
       });
     }
     recursiveFunc(formToInvestigate);
-    return invalidControls;
   }
 }
