+++ libs/movie/src/lib/movie/pages/summary/promotional/media-files/media-files.component.scss
@@ -5,4 +5,5 @@ strong {
 
 h6 {
   margin: 0;
+  opacity: 0.5;
 }
