--- libs/movie/src/lib/movie/pages/summary/summary.component.scss
@@ -1,35 +1,0 @@
-li {
-  margin-bottom: 4px;
-}
-
-h3 {
-  margin-bottom: 24px;
-}
-
-.mat-icon {
-  vertical-align: middle;
-  margin-right: 8px;
-}
-
-mat-divider {
-  border-top-width: 2px;
-  margin-top: 36px;
-  margin-bottom: 36px;
-}
-
-button.submit {
-  margin: auto;
-}
-
-.mat-body {
-  margin-bottom: 36px;
-
-  & span {
-    display: block;
-  }
-
-}
-
-article {
-  margin-left: 16px;
-}
