+++ libs/movie/src/lib/movie/pages/summary/synopsis/synopsis.component.scss
@@ -1,3 +1,4 @@
 h6 {
   margin: 0;
+  opacity: 0.5;
 }
